Using the ng-content directive.
By default, in Angular 5, everything placed within the opening and closing tag of a component is lost. The ng-content directive overrides this, outputting everything found within the opening and closing tags where it's placed.

This is useful for reusable components! In this case, it made the code cleaner by allowing all the control code to be placed in a single file.