This lecture focused on the behavior of lifecycle hooks. Basically, when a component is created, Angular cycles through a few different phases, which lifecycle hooks allow us to take advantage of!
The first phase is ngOnChanges, which executes at the start, as well as any time a bound input property changes.

The second phase is ngOnInit, which is only called once the component is initialized. This isn't the same as creation! It isn't added to the DOM yet, just read. Runs after the constructor.

The third is ngDoCheck, which is called any time Angular's change detection runs - even if nothing changed!

The fourth is ngAfterContentInit, which is called after the content within the component (eg, ng-content, from the directive we used before) is projected into the view.

The fifth is ngAfterContentChecked, which is called after change detection runs on the content.

The sixth is ngAfterViewInit after the component's view (and child views) have been initialized. IE, after they've been rendered in the DOM.

Finally, if you destroy a component, ngOnDestroy is called.