It turned out in the last lesson the service didn't appear to update when called in the child components! Why?
Angular uses 'hierarchical injection'. What this means is that when a service is injected (IE, included in the 'providers' array in the @Component decorator), the same instance becomes available to all child components.

When a service is injected in AppModule, it's available across the entire application. When it's injected into AppComponent, it's available for all components, but NOT other services. When added to any other component, it's available to the component and children.

Crucially, when you reinject the service into the child of a component that's already been injected, they do NOT share the same instance. The child (and its children!) get a new instance of the service.

Therefore, by removing AccountsService from the new-account and account components 'providers' array, but still creating them in the constructor, we allow them to access the same instance of AccountsService as generated in the App component.