Creating custom validators, which unlike most form stuff, is a little tricky!
If you return 'true', the validator throws the name/key as an error. The catch is that you don't return false if it's valid. Instead, you return nothing or null.

The other trick is that when passing the function from the component, the context will be lost when the function is used by the validator. To avoid this, you gotta bind it to 'this'.